requirements.txt
Flask==2.3.2
gspread==5.8.0
google-auth==2.21.0
